day02_data_models/pyproject.toml
[project]
name = "day02_data_models"
version = "0.1.0"
dependencies = [
    "mimesis>=18.0.0",
    "neo4j>=5.28.2",
    "pgcli>=4.3.0",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.14.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]
